+++ frontend/app/src/people/widgetViews/postBounty/__tests__/PostModal.spec.tsx
@@ -3,10 +3,10 @@ import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import nock from 'nock';
 import React from 'react';
-import { setupStore } from '../../../__test__/__mockData__/setupStore';
-import { user } from '../../../__test__/__mockData__/user';
-import { mockUsehistory } from '../../../__test__/__mockFn__/useHistory';
-import { PostModal } from './PostModal';
+import { setupStore } from '../../../../__test__/__mockData__/setupStore';
+import { user } from '../../../../__test__/__mockData__/user';
+import { mockUsehistory } from '../../../../__test__/__mockFn__/useHistory';
+import { PostModal } from '../PostModal';
 
 beforeAll(() => {
   nock.disableNetConnect();
